값을 내는 자리에서 값을 확인하고, 그 자리에서 저장한다
재화를 GetItem() 말고 robin 에서 직접 빼는 자리가 여섯 군데 있었다.
훑어보니 세 종류로 갈렸다.

살아 있던 것 - 상자 구매
--------------------------------------------------------------------------
살 수 있는지를 버튼을 그릴 때만 봤다. 값을 내는 자리에서는 안 봤다.
저장도 없어서, 골드는 나갔는데 상자는 못 받은 상태가 될 수 있었다.

GetBoxPrice() 에는 default 가 없었다. 목록에 없는 상자가 오면 쓰레기값이
그대로 값이 된다. int 라서 성이 46 개가 되면 값이 음수가 되고, 음수를
빼면 골드가 늘어난다. 지금 성이 32 개라 여유가 두 배뿐이었다.
long long 으로 바꾸고 못 파는 상자는 -1 로 답하게 했다.

같은 병이었는데 못 보던 것 - 장비 강화와 진화
--------------------------------------------------------------------------
버튼을 그릴 때 robin.hammer >= 값 을 보고, 실제 차감은 60 프레임 뒤
연출 프레임에서 확인 없이 했다. 연출이 모달이라 그 사이 재화가 줄 일은
없어 보이지만, 적혀 있지도 지켜지지도 않는 약속에 기대고 있었다.
차감하는 자리에서 다시 보고, 모자라면 재화도 효과도 둘 다 안 건드린다.

동료 강화는 CanCrewLevelUp() 이 소비 시점에 골드까지 본다. 원래 맞았다.

죽어 있던 것 - 스킬 강화, 아레나 부활
--------------------------------------------------------------------------
버튼이 안 붙어 있어 여기까지 못 온다. 그래서 아무도 손해를 안 봤다.

스킬 강화는 골드만 가져가고 아무 일도 안 했다. skillLv 는 어디서도 안
오르고, 설정하는 winUpgradeFrame 은 아무도 안 읽는다. 미완성 기능이 그대로
남아 있었다. 버튼을 붙이는 날 사고가 나므로 재화를 안 건드리게 했다.

연출이 값을 정하던 자리들
--------------------------------------------------------------------------
골드를 바에서 읽어 지급하는 곳이 셋 있었다. 그런데 BAR_COIN 에 0 아닌
값을 넣는 살아 있는 코드가 없다 - 바탕값을 넣던 줄이 둘 다 주석이고,
나머지는 자기 자신에 비율을 곱하는 것뿐이라 0 에 곱해 0 이다. 지금은
정확히 0 원을 준다. 그래도 고쳤다. 그 주석을 되살리는 날, 자르는 기준이
robin.gold 가 아니라 굴러가는 골드 바인 것과 저장이 없는 것이 그대로면
조용히 틀린다.

그리고 배틀 입장료가 한 번도 걷힌 적이 없었다. AddBar 하나가 전부였고,
바는 robin.heart 를 안 건드린다. 화면에서만 줄었다가 다음 동기화 때
되돌아왔다. GetStageAdmissionHeart() 를 쓰는 자리가 그 AddBar 하나였다.

+++ Classes/Func_Menu.h
@@ -132,7 +132,7 @@ int GetBoxCurrency(int, int);
 int GetBoxGold(int boxType);
 int GetBoxQuestItem(int boxType);
 int GetBoxHeartItem(int boxType);
-int GetBoxPrice(int, int);
+long long GetBoxPrice(int, int);
 
 int GetItemCategoryCnt(int category);
 int GetScrollDy(int);
